requirements.txt
torch==1.13.1
torchvision==0.14.1
torchtext==0.14.1
torchaudio==0.13.1
torchdata==0.5.1
torchsummaryX==1.3.0
pandas==1.5.2
wandb
python-Levenshtein
ctcdecode
numpy
scikit-learn
tqdm